record page pt-1: desciption and metadata
