+++ packages/web-app/src/lib/components/card/card.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+</script>
+
+<div class="flex flex-col space-y-7 p-6 my-2 bg-custom-5">
+  <slot />
+</div>
