+++ packages/web-app/src/lib/components/record/i18n/en/translations.json
@@ -1,0 +1,46 @@
+{
+  "aboutThis": "About this Dataset",
+  "accessAllTime": "All time access",
+  "accessLast30": "Access in last 30 days",
+  "address": "Address",
+  "addToMap": "Add to Map",
+  "addToMapDescription": "View the data in depth by adding it to a map.",
+  "advancedMetaData": "Advanced MetaData",
+  "contactData": "Contact Data",
+  "dataResources": "Data Resources",
+  "dateCreated": "Date Created",
+  "datePublished": "Date Published",
+  "description": "Description",
+  "downloadGeoCore": "Download geoCore",
+  "east": "East",
+  "email": "Email",
+  "fax": "Fax",
+  "format": "Format",
+  "id": "ID",
+  "individualName": "Individual Name",
+  "keywords": "keywords",
+  "languages": "Languages",
+  "maintenance": "Maintenance",
+  "metadata": "Metadata",
+  "metadataDescription": "Our metadata is stored in the geoCore format. A geojson containing all the metadata you see here.",
+  "name": "Name",
+  "north": "North",
+  "organization": "Organization",
+  "role": "Role",
+  "similarRecords": "Similar Records",
+  "sources": "Source(s)",
+  "south": "South",
+  "spatialRepresentation": "Spatial Representation",
+  "status": "Status",
+  "telephone": "Telephone",
+  "temporalCoverage": "Temporal Coverage",
+  "topicCategory": "Topic Category",
+  "type": "Type",
+  "useLimitations": "Use Limitations",
+  "viewHNAP": "View HNAP Record",
+  "viewLess": "View Less",
+  "viewMore": "View More",
+  "viewOnMap": "View on Map",
+  "website": "Website",
+  "west": "West"
+}
