+++ packages/web-app/src/lib/components/record/i18n/fr/translations.json
@@ -1,0 +1,46 @@
+{
+  "aboutThis": "À propos de cet ensemble de données",
+  "accessAllTime": "Accès à tout moment",
+  "accessLast30": "Accès dans les 30 derniers jours",
+  "address": "Adresse",
+  "addToMap": "Ajouter à la carte",
+  "addToMapDescription": "Affichez les données en profondeur en les ajoutant à une carte.",
+  "advancedMetaData": "Métadonnées avancées",
+  "contactData": "Données de contact",
+  "dataResources": "Ressources de données",
+  "dateCreated": "Date créée",
+  "datePublished": "Date publiée",
+  "description": "Description",
+  "downloadGeoCore": "Télécharger geoCore",
+  "east": "Est",
+  "email": "Courriel",
+  "fax": "Télécopieur",
+  "format": "Format",
+  "id": "Identifiant",
+  "individualName": "Nom de la personne",
+  "keywords": "Mots clés",
+  "languages": "Langue",
+  "maintenance": "Fréquence de mise à jour",
+  "metadata": "Métadonnées",
+  "metadataDescription": "Nos métadonnées sont stockées au format geoCore. Un geojson contenant toutes les métadonnées que vous voyez ici.",
+  "name": "Nom",
+  "north": "Nord",
+  "organization": "Organisation",
+  "role": "Rôle",
+  "similarRecords": "Dossiers similaires",
+  "sources": "Source(s)",
+  "south": "Sud",
+  "spatialRepresentation": "Représentation Spatiale",
+  "status": "État",
+  "telephone": "Téléphone",
+  "temporalCoverage": "Couverture temporelle",
+  "topicCategory": "Catégorie thématique",
+  "type": "Type",
+  "useLimitations": "Limites d'utilisation",
+  "viewHNAP": "Enregistrement HNAP",
+  "viewLess": "Montrer Moins",
+  "viewMore": "Voir Plus",
+  "viewOnMap": "Voir sur la carte",
+  "website": "web",
+  "west": "Ouest"
+}
