+++ packages/web-app/src/lib/components/record_old/i18n/fr.json
