+++ packages/web-app/src/lib/components/record_old/modal/info-modal.svelte
