+++ packages/web-app/tailwind.config.js
@@ -21,7 +21,8 @@ export default {
 					14: '#2C3034',
 					15: '#212529',
 					16: '#535AA4',
-					17: '#e0e0e0'
+					17: '#e0e0e0',
+					18: '#00000029'
 				}
 			},
 			spacing: {
